requirements.txt
streamlit>=1.33
tensorflow>=2.20,<2.22
numpy>=1.26,<3
Pillow>=10.0
